+++ camera-qr/src/commonMain/kotlin/tech/kotlinlang/camera/qr/QrCodeImageAnalyser.kt
@@ -1,0 +1,5 @@
+package tech.kotlinlang.camera.qr
+
+import tech.kotlinlang.camera.analyser.ImageAnalyser
+
+expect class QrCodeImageAnalyser: ImageAnalyser<String>