+++ camera/src/androidMain/kotlin/tech/kotlinlang/camera/analyser/AndroidImageAnalyser.kt
@@ -1,0 +1,8 @@
+package tech.kotlinlang.camera.analyser
+
+import androidx.camera.core.ImageProxy
+
+interface AndroidImageAnalyser<T>: ImageAnalyser<T> {
+    fun setImageProxy(imageProxy: ImageProxy)
+    fun setStopListener(callback: () -> Unit)
+}