+++ camera/src/androidMain/kotlin/tech/kotlinlang/camera/analyser/EmptyImageAnalyser.kt
@@ -1,0 +1,9 @@
+package tech.kotlinlang.camera.analyser
+
+import androidx.camera.core.ImageProxy
+
+actual class EmptyImageAnalyser: AndroidImageAnalyser<Unit> {
+    override fun setImageProxy(imageProxy: ImageProxy) {}
+    override fun setListener(callback: (Unit) -> Unit) {}
+    override fun setStopListener(callback: () -> Unit) {}
+}