+++ camera/src/commonMain/kotlin/tech/kotlinlang/camera/CameraHelper.kt
@@ -2,8 +2,9 @@ package tech.kotlinlang.camera
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import tech.kotlinlang.camera.analyser.ImageAnalyser
 
 interface CameraHelper {
     @Composable
-    fun CameraContent(modifier: Modifier)
+    fun <T> CameraContent(modifier: Modifier, imageAnalyser: ImageAnalyser<T>)
 }