+++ camera/src/commonMain/kotlin/tech/kotlinlang/camera/analyser/EmptyImageAnalyser.kt
@@ -1,0 +1,3 @@
+package tech.kotlinlang.camera.analyser
+
+expect class EmptyImageAnalyser: ImageAnalyser<Unit>