+++ camera/src/commonMain/kotlin/tech/kotlinlang/camera/analyser/ImageAnalyser.kt
@@ -1,0 +1,5 @@
+package tech.kotlinlang.camera.analyser
+
+interface ImageAnalyser<T> {
+    fun setListener(callback: (T) -> Unit)
+}