+++ camera/src/iosMain/kotlin/tech/kotlinlang/camera/analyser/EmptyImageAnalyser.kt
@@ -1,0 +1,12 @@
+package tech.kotlinlang.camera.analyser
+
+import platform.AVFoundation.AVCaptureOutput
+
+actual class EmptyImageAnalyser : IosImageAnalyser<Unit> {
+    override fun provideAVCaptureOutput(): AVCaptureOutput? {
+        return null
+    }
+
+    override fun initializeAVCaptureOutput(output: AVCaptureOutput) {}
+    override fun setListener(callback: (Unit) -> Unit) {}
+}