+++ camera/src/iosMain/kotlin/tech/kotlinlang/camera/analyser/IosImageAnalyser.kt
@@ -1,0 +1,8 @@
+package tech.kotlinlang.camera.analyser
+
+import platform.AVFoundation.AVCaptureOutput
+
+interface IosImageAnalyser<T>: ImageAnalyser<T> {
+    fun provideAVCaptureOutput(): AVCaptureOutput?
+    fun initializeAVCaptureOutput(output: AVCaptureOutput)
+}